Fix integer portability
On some architectures, `int` type is 16 bits (C99's minimum size).
Because it cannot store numbers higher than 32767, two things break:
- keyboard commands, like `UP_ARROW` (63232)
- `rand_range()` is called with numbers as high as 60000

For those, use `long` instead.

+++ src/brogue/Rogue.h
@@ -2589,7 +2589,7 @@ extern "C" {
     void gameOver(char *killedBy, boolean useCustomPhrasing);
     void victory(boolean superVictory);
     void enableEasyMode();
-    int rand_range(int lowerBound, int upperBound);
+    long rand_range(long lowerBound, long upperBound);
     unsigned long seedRandomGenerator(unsigned long seed);
     short randClumpedRange(short lowerBound, short upperBound, short clumpFactor);
     short randClump(randomRange theRange);
